fix: Refactor packages names
