--- lib/src/main/kotlin/treeLib/RBtree/RBStruct.kt
@@ -1,9 +1,0 @@
-package treeLib.RBtree
-
-import treeLib.AbstractTree.Weighted.WeightedTreeStruct
-
-class RBStruct <Pack : Comparable<Pack>> : WeightedTreeStruct<Pack, RBNode<Pack>, RBStateContainer<Pack>, RBBalancer<Pack>>(){
-    override var root: RBNode<Pack>? = null
-    override val balancer = RBBalancer<Pack>(root)
-
-}
