+++ lib/src/main/kotlin/treelib/abstractTree/Node.kt
@@ -1,4 +1,4 @@
-package treeLib.AbstractTree
+package treelib.abstractTree
 
 interface Node<Pack : Comparable<Pack>, SubNode : Node<Pack, SubNode>>{
     var value: Pack
