+++ lib/src/main/kotlin/treelib/abstractTree/StateContainer.kt
@@ -1,4 +1,4 @@
-package treeLib.AbstractTree
+package treelib.abstractTree
 
 interface StateContainer<V: Comparable<V>, NodeType: Node<V, NodeType>> {
     val contentNode: NodeType
