+++ lib/src/main/kotlin/treelib/abstractTree/balanced/BalancerParent.kt
@@ -1,7 +1,7 @@
-package treeLib.AbstractTree.Weighted
+package treelib.abstractTree.balanced
 
-import treeLib.AbstractTree.NodeParent
-import treeLib.AbstractTree.StateContainer
+import treelib.abstractTree.NodeParent
+import treelib.abstractTree.StateContainer
 
 abstract class BalancerParent<Pack : Comparable<Pack>, NodeType : NodeParent<Pack, NodeType>, StateContainerType: StateContainer<Pack, NodeType>>: Balancer<Pack, NodeType, StateContainerType> {
     override fun rightRotate(currentNode: NodeType): NodeType {
