+++ lib/src/main/kotlin/treelib/rbTree/RBTree.kt
@@ -1,8 +1,8 @@
-package treeLib.RBtree
+package treelib.rbTree
 
-import treeLib.AbstractTree.Tree
+import treelib.abstractTree.Tree
 
-import treeLib.Single_Objects.Container
+import treelib.singleObjects.Container
 
 class RBTree<Key : Comparable<Key>, Value> : Tree<Key, Value, RBNode<Container<Key, Value>>>() {
     override val treeStruct: RBStruct<Container<Key, Value>> = RBStruct()
