+++ lib/src/main/kotlin/treelib/singleObjects/Container.kt
@@ -1,4 +1,4 @@
-package treeLib.Single_Objects
+package treelib.singleObjects
 
 class Container<K : Comparable<K>, V>(private val pair: Pair<K, V?>) : Comparable<Container<K, V>> {
     val key = pair.first
